# Remove memory churn in compute_affine_map