@jwpeterson 2016-08-25
Seems like a good fix to me, this function should never be called by multiple threads on the same object, so there should be no issues with making that a class variable...