@roystgnr 2016-08-26
Awesome catch, but can I ask some stupid questions?

Does the same problem apply when considering vector declarations inside vs outside loops, or is the compiler smart enough to optimize away the implied redundant free-then-malloc-again-right-away?  My coding style is to declare variables in the innermost scope possible, but if that's going to be slowing us down by a factor of 3 on many-core CPUs then it's time to rethink that style.

Could we have worked around the problem by instead using a non-default allocator or a sufficiently large array or std::array?  Something that uses ~450 bytes of memory on the stack would end up never calling heap management at all, right?  But then we'd either need our helper functions to only take vectors specialized with that allocator or we'd need them to take pointer instead of vector arguments.

Are the standard malloc and free sufficiently thread-aware?  I'd have expected an order of magnitude less slowdown from memory management as simple as "chop up virtual memory into thread-specific spaces, don't edit anyone else's space during malloc, save the slow code paths for when one thread is asked to free another thread's malloc output".