@permcody 2016-08-26
On Fri, Aug 26, 2016 at 8:35 AM roystgnr <notifications@github.com> wrote:

> Awesome catch, but can I ask some stupid questions?
>
> Does the same problem apply when considering vector declarations inside vs
> outside loops, or is the compiler smart enough to optimize away the implied
> redundant free-then-malloc-again-right-away? My coding style is to declare
> variables in the innermost scope possible, but if that's going to be
> slowing us down by a factor of 3 on many-core CPUs then it's time to
> rethink that style.
>
> Could we have worked around the problem by instead using a non-default
> allocator or a sufficiently large array or std::array? Something that uses
> ~450 bytes of memory on the stack would end up never calling heap
> management at all, right? But then we'd either need our helper functions to
> only take vectors specialized with that allocator or we'd need them to take
> pointer instead of vector arguments.
>

I'm sure that any of those options would work but now you are writing
slightly more complex and slightly less elegant code. Any trip to the
memory manager in threaded or non-threaded regions is always slow,
sometimes due to the necessary locking to make it thread safe. As a first
take on any optimization in a C++ algorithm we should probably strive to
reduce trips to the allocator whenever possible even if it means wasting a
little space. Also, I like the idea of creating more stack variables and
using the heap a little less again, that's got to help a ton in threaded
code. My guess is, it'll remove some of the "False Sharing" crap Derek is
seeing now.


> Are the standard malloc and free sufficiently thread-aware? I'd have
> expected an order of magnitude less slowdown from memory management as
> simple as "chop up virtual memory into thread-specific spaces, don't edit
> anyone else's space during malloc, save the slow code paths for when one
> thread is asked to free another thread's malloc output".
>

I'm believe that "new" and "delete" are typically thread-safe but I'm
pretty sure that standard definitions of malloc and free are not since they
are ANSI-C.


> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/libMesh/libmesh/pull/1072#issuecomment-242752888>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AC5XIC9OR3dh_LxCpYSc-U35udAdXimVks5qjvnNgaJpZM4JsmEs>
> .
>