@friedmud 2016-08-26
@permcody got this right... whenever possible we need to remove ANY amount of allocation / deallocation in compute loops (regardless of threading).

Memory churn slows things down even in single-threaded applications but with threading it hammers you (as you can see here).

>  is the compiler smart enough to optimize away the implied redundant free-then-malloc-again-right-away

As far as I know: no... not for a std::vector

> Could we have worked around the problem by instead using a non-default allocator or a sufficiently large array or std::array? Something that uses ~450 bytes of memory on the stack would end up never calling heap management at all, right? But then we'd either need our helper functions to only take vectors specialized with that allocator or we'd need them to take pointer instead of vector arguments.

I believe so (for the large stack array): as long as it's not too large.  But I agree with @permcody that this is inelegant and overly complex.

As for allocators... "scalable allocators" _are_ a thing.  But they are complex, and as far as I know there are none built in.  TBB comes with some you can use: https://www.threadingbuildingblocks.org/tutorial-intel-tbb-scalable-memory-allocator

But, again, that's overly complex for just a simple little vector like this.  Use of these should be relegated to the "only if you really have to" bin.

(Note: That page also talks about "False Sharing" a bit... which is ANOTHER threading problem I've been fighting lately...)

> I'm believe that "new" and "delete" are typically thread-safe but I'm
pretty sure that standard definitions of malloc and free are not since they
are ANSI-C.

They are all thread safe (even `malloc`) (if they weren't we would be a SHIT TON of trouble!)... and is one of the major reasons why they're slow.  Essentially they all use some form of locking to maintain thread safety... that is a global lock... so every call to `new` is contending for the same lock.  If a couple of calls to `new` are in compute sensitive areas then this can effectively serialize your code!